Print basic info per generation

Print basic info per generation

callback_base_report(
  userdir = "",
  generation,
  pop_pool,
  pareto,
  prob_matrix,
  current_time
)

Arguments

userdir
the default directory used by `galgo()` to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with base_report_callback assigned to the report_callback
# hook-point
galgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
report_callback = callback_base_report
)

#> Using CPU for computing pearson distance

#> Generation 1 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.7  2 0.10888047  14.49866        1       Inf
#> result.9  5 0.05056947 185.49095        1 0.7214049
#> result.10 2 0.06498447 160.66696        1 1.3679277
#> result.14 7 0.02128337 224.89984        1       Inf

#> Generation 2 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.7  2 0.10888047  14.49866        1       Inf
#> result.14 7 0.02128337 224.89984        1       Inf
#> result.9  5 0.05056947 185.49095        1 0.7175245
#> result.13 2 0.09296446  76.69576        1 0.6351903
#> result.8  2 0.07819196  87.59060        1 0.6134396
#> result.9  3 0.06559908 151.89572        1 0.4686648
#> result.10 2 0.06498447 160.66696        1 0.3007531

#> Generation 3 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.00939784 294.48435        1       Inf
#>            7 0.02889464 237.78186        1 0.6442515
#> result.13  2 0.09296446  76.69576        1 0.5460982
#> result.8   2 0.07819196  87.59060        1 0.5224424
#>            5 0.03951537 192.15347        1 0.3880625
#> result.9   3 0.06559908 151.89572        1 0.3823506
#> result.9   5 0.05056947 185.49095        1 0.3499488
#> result.10  2 0.06498447 160.66696        1 0.2596578

#> Generation 4 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.00939784 294.48435        1       Inf
#>            7 0.02889464 237.78186        1 0.6193555
#>            2 0.08554100 132.75042        1 0.4358089
#>            2 0.10800606  61.41323        1 0.3975359
#>            5 0.03951537 192.15347        1 0.3731564
#> result.9   5 0.05056947 185.49095        1 0.3366572
#> result.9   3 0.06559908 151.89572        1 0.2802297
#>            2 0.09556422  96.13645        1 0.2540836
#> result.10  2 0.06498447 160.66696        1 0.2496973
#>            2 0.09516738 102.11118        1 0.2148175

#> Generation 5 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.00939784 294.48435        1       Inf
#>            7 0.02889464 237.78186        1 0.6231472
#>            2 0.08554100 132.75042        1 0.4385594
#>            2 0.10800606  61.41323        1 0.3999103
#>            5 0.03951537 192.15347        1 0.3754756
#> result.9   5 0.05056947 185.49095        1 0.3388079
#> result.9   3 0.06559908 151.89572        1 0.2820135
#>            2 0.09556422  96.13645        1 0.2556438
#> result.10  2 0.06498447 160.66696        1 0.2512543
#>            2 0.09516738 102.11118        1 0.2161284

#> An object of class "galgo.Obj"
#> Slot "Solutions":
#>           205766_at 38241_at 207078_at 210843_s_at 200957_s_at 219627_at
#> result.7          0        0         1           0           0         1
#>                   1        1         0           1           1         1
#>                   0        0         1           0           0         1
#>                   1        0         0           0           1         1
#>                   0        0         0           0           0         1
#>                   1        1         1           1           1         1
#> result.9          1        0         0           0           1         0
#> result.9          1        1         1           1           1         1
#>                   0        0         1           1           1         1
#> result.10         1        1         1           1           1         1
#>                   0        0         1           0           0         1
#> result.13         1        0         1           1           1         1
#>                   1        1         0           1           1         1
#>                   1        1         0           1           1         1
#>                   1        1         1           1           1         1
#>           215890_at 218495_at 212079_s_at 202717_s_at 221636_s_at 213443_at
#> result.7          0         1           0           0           0         0
#>                   0         0           1           1           1         1
#>                   0         1           1           0           0         0
#>                   0         0           0           1           1         0
#>                   0         1           0           0           0         0
#>                   1         1           1           1           1         1
#> result.9          0         0           0           0           0         0
#> result.9          1         1           1           1           1         1
#>                   0         1           0           0           1         0
#> result.10         1         1           1           1           1         1
#>                   0         1           0           0           0         0
#> result.13         0         1           0           0           1         0
#>                   0         0           1           1           1         1
#>                   0         0           1           1           1         1
#>                   0         1           1           1           1         1
#>           204646_at 219635_at 212306_at 216050_at 213462_at 216458_at 203845_at
#> result.7          0         1         0         0         0         1         1
#>                   1         1         1         1         1         1         1
#>                   0         0         1         0         0         1         0
#>                   0         1         1         0         1         1         0
#>                   0         1         0         0         0         1         1
#>                   1         1         1         1         1         1         1
#> result.9          0         1         1         0         1         1         0
#> result.9          1         1         1         1         1         1         1
#>                   0         1         0         0         0         1         1
#> result.10         1         1         1         1         1         1         1
#>                   0         1         0         0         0         1         1
#> result.13         0         1         0         1         0         1         1
#>                   1         1         1         1         1         1         1
#>                   1         1         1         1         1         1         1
#>                   0         1         1         1         0         1         1
#>           202453_s_at 213453_x_at 212975_at 202519_at 215409_at 220352_x_at
#> result.7            0           0         0         0         1           0
#>                     1           0         1         0         1           1
#>                     1           1         0         1         1           1
#>                     0           0         0         0         0           1
#>                     0           0         0         0         1           1
#>                     1           1         0         1         1           1
#> result.9            0           0         0         0         0           1
#> result.9            0           1         1         0         1           1
#>                     0           1         0         0         1           1
#> result.10           1           1         1         1         1           1
#>                     0           0         0         0         1           0
#> result.13           0           1         0         0         1           1
#>                     1           0         1         0         1           1
#>                     1           0         1         0         1           1
#>                     0           1         1         1         1           1
#>           209451_at 214150_x_at 212434_at 214840_at 212423_at 209798_at
#> result.7          0           0         0         0         1         0
#>                   0           0         1         0         0         0
#>                   1           0         1         1         1         0
#>                   1           0         1         0         1         0
#>                   0           0         0         0         1         0
#>                   0           1         1         1         1         1
#> result.9          1           1         0         1         1         0
#> result.9          1           1         1         0         1         0
#>                   0           0         0         0         1         0
#> result.10         0           1         1         1         1         1
#>                   0           0         0         0         1         0
#> result.13         0           0         0         0         1         0
#>                   0           0         1         0         0         0
#>                   0           0         1         0         0         0
#>                   1           1         1         1         1         0
#>           211061_s_at 206768_at 218042_at 218309_at 221297_at 212062_at
#> result.7            0         0         0         0         0         1
#>                     1         1         1         1         1         1
#>                     1         1         0         0         1         0
#>                     0         0         1         0         1         1
#>                     1         0         0         0         0         1
#>                     0         1         1         1         1         1
#> result.9            0         0         0         0         1         1
#> result.9            1         1         1         1         1         1
#>                     1         0         1         0         1         1
#> result.10           1         1         1         1         1         1
#>                     0         0         0         0         0         1
#> result.13           1         0         1         0         1         1
#>                     1         1         1         1         1         1
#>                     1         1         1         1         1         1
#>                     1         1         1         1         1         0
#>           202542_s_at 218941_at 214003_x_at 214608_s_at 204748_at 221953_s_at
#> result.7            1         0           0           0         0           0
#>                     1         0           1           1         0           0
#>                     0         1           0           1         1           0
#>                     0         0           0           1         1           0
#>                     1         0           0           0         0           0
#>                     1         1           1           1         0           0
#> result.9            0         0           0           1         1           0
#> result.9            1         1           1           1         1           0
#>                     1         0           0           1         0           0
#> result.10           1         1           1           1         1           0
#>                     1         0           0           0         0           0
#> result.13           1         0           0           1         0           0
#>                     1         0           1           1         0           0
#>                     1         0           1           1         0           0
#>                     1         1           1           1         1           0
#>           218580_x_at 205064_at 44563_at 206783_at 213170_at 203257_s_at
#> result.7            0         0        0         0         0           0
#>                     1         0        1         0         1           1
#>                     0         1        0         0         1           1
#>                     0         0        1         1         1           1
#>                     0         0        0         0         0           0
#>                     1         1        1         1         1           1
#> result.9            0         0        1         1         1           0
#> result.9            1         1        1         1         1           1
#>                     0         0        1         0         0           0
#> result.10           1         1        1         1         1           1
#>                     0         0        0         0         0           0
#> result.13           0         0        1         0         0           0
#>                     1         0        1         0         1           1
#>                     1         0        1         0         1           1
#>                     0         1        1         1         1           1
#>           203397_s_at 218895_at 202312_s_at 201787_at 204400_at 202302_s_at
#> result.7            0         1           1         0         0           0
#>                     1         1           1         1         0           1
#>                     1         0           0         0         1           1
#>                     1         0           0         0         1           0
#>                     0         1           1         0         0           0
#>                     1         1           1         1         1           1
#> result.9            1         0           0         0         1           0
#> result.9            1         1           1         1         1           1
#>                     0         1           1         0         0           0
#> result.10           1         1           1         1         1           1
#>                     0         1           1         0         0           0
#> result.13           0         1           1         0         0           1
#>                     1         1           1         1         0           1
#>                     1         1           1         1         0           1
#>                     1         1           1         0         1           1
#>           207070_at 217765_at 219645_at 214593_at 211868_x_at 204315_s_at
#> result.7          0         0         0         0           1           0
#>                   0         0         1         0           1           1
#>                   0         0         0         0           1           1
#>                   0         1         0         0           1           1
#>                   0         0         0         0           1           0
#>                   1         1         0         1           1           1
#> result.9          0         1         0         0           1           0
#> result.9          1         0         1         0           1           1
#>                   0         0         0         0           0           1
#> result.10         1         1         1         1           1           1
#>                   0         0         0         0           1           0
#> result.13         0         0         0         0           0           1
#>                   0         0         1         0           1           1
#>                   0         0         1         0           1           1
#>                   1         0         1         0           1           1
#>           210288_at 201689_s_at 214930_at 216444_at 203557_s_at 39835_at
#> result.7          0           1         0         1           0        0
#>                   1           1         1         1           1        1
#>                   1           1         1         1           1        0
#>                   0           0         0         0           1        0
#>                   0           1         0         1           0        0
#>                   1           1         1         1           1        1
#> result.9          0           0         0         0           1        0
#> result.9          1           1         1         1           1        1
#>                   1           1         0         1           0        0
#> result.10         1           1         1         1           1        1
#>                   0           1         0         1           0        0
#> result.13         1           1         0         1           1        0
#>                   1           1         1         1           1        1
#>                   1           1         1         1           1        1
#>                   1           1         1         1           1        1
#>           202305_s_at 220670_at 221219_s_at 214941_s_at 211665_s_at 200974_at
#> result.7            1         0           0           0           0         0
#>                     1         1           0           1           1         1
#>                     0         1           0           0           1         1
#>                     1         1           1           1           0         1
#>                     1         1           0           1           0         0
#>                     1         1           1           1           1         1
#> result.9            1         1           1           1           0         1
#> result.9            1         1           1           1           1         1
#>                     1         0           1           0           0         0
#> result.10           1         1           1           1           1         1
#>                     1         0           0           0           0         0
#> result.13           1         0           1           0           0         0
#>                     1         1           0           1           1         1
#>                     1         1           0           1           1         1
#>                     1         1           1           1           1         1
#>           211066_x_at 201245_s_at 42361_g_at 214304_x_at 201225_s_at 1494_f_at
#> result.7            1           1          0           1           0         0
#>                     0           1          1           1           1         1
#>                     0           1          0           1           0         1
#>                     0           0          1           1           1         0
#>                     1           1          0           1           0         0
#>                     1           1          1           0           1         1
#> result.9            0           0          1           0           0         0
#> result.9            1           1          1           1           1         1
#>                     1           1          0           1           1         1
#> result.10           1           1          1           1           1         1
#>                     1           1          0           1           0         0
#> result.13           1           1          0           1           1         1
#>                     0           1          1           1           1         1
#>                     0           1          1           1           1         1
#>                     1           1          1           1           1         1
#>           218653_at 206010_at 209287_s_at 204857_at 217214_s_at 216919_at
#> result.7          0         0           1         1           0         0
#>                   1         0           1         0           1         0
#>                   0         0           0         1           1         0
#>                   1         0           1         0           1         0
#>                   0         0           1         1           0         0
#>                   1         1           1         1           1         1
#> result.9          1         0           1         0           1         0
#> result.9          1         1           1         1           1         0
#>                   1         0           1         1           1         0
#> result.10         1         1           1         1           1         1
#>                   0         0           1         1           0         0
#> result.13         1         0           1         1           1         0
#>                   1         0           1         0           1         0
#>                   1         0           1         0           1         0
#>                   1         0           1         1           1         0
#>           218306_s_at 208328_s_at 213379_at 211370_s_at 211263_s_at 203906_at
#> result.7            0           0         0           0           1         0
#>                     1           1         1           1           1         0
#>                     1           1         0           1           1         1
#>                     0           1         1           1           0         0
#>                     1           0         0           0           1         0
#>                     1           1         1           1           1         0
#> result.9            0           0         0           1           0         1
#> result.9            1           0         1           1           1         0
#>                     0           0         1           1           1         0
#> result.10           1           1         1           1           1         1
#>                     0           0         0           0           1         0
#> result.13           1           0         1           1           1         0
#>                     1           1         1           1           1         0
#>                     1           1         1           1           1         0
#>                     1           0         1           1           1         0
#>           205524_s_at 212900_at 216668_at 203398_s_at 205106_at 207812_s_at
#> result.7            0         0         0           0         0           0
#>                     1         1         0           1         1           1
#>                     0         1         1           0         1           0
#>                     0         1         1           1         1           0
#>                     0         0         0           1         0           0
#>                     1         1         1           1         0           1
#> result.9            0         1         1           1         1           0
#> result.9            0         1         1           1         1           1
#>                     0         1         1           1         1           0
#> result.10           1         1         1           1         1           1
#>                     0         0         0           0         0           0
#> result.13           0         1         1           1         1           0
#>                     1         1         0           1         1           1
#>                     1         1         0           1         1           1
#>                     0         1         1           1         1           1
#>           211921_x_at 202926_at 206606_at  k                        rnkIndex
#> result.7            0         1         0  2  0.108880468  14.49866        1
#>                     1         1         0 10  0.009397840 294.48435        1
#>                     0         1         1  7  0.028894640 237.78186        1
#>                     0         0         1  2  0.085541004 132.75042        1
#>                     0         1         0  2  0.108006062  61.41323        1
#>                     1         1         1  5  0.039515369 192.15347        1
#> result.9            0         0         1  5  0.050569465 185.49095        1
#> result.9            1         0         1  3  0.065599078 151.89572        1
#>                     0         1         0  2  0.095564224  96.13645        1
#> result.10           1         1         1  2  0.064984472 160.66696        1
#>                     0         1         0  2  0.095167382 102.11118        1
#> result.13           0         1         0  2  0.092964459  76.69576        2
#>                     1         1         0 10 -0.001285935 250.12797        2
#>                     1         1         0 10 -0.001285935 250.12797        2
#>                     1         0         1  4  0.051347347 138.06628        2
#>               CrowD
#> result.7        Inf
#>                 Inf
#>           0.6231472
#>           0.4385594
#>           0.3999103
#>           0.3754756
#> result.9  0.3388079
#> result.9  0.2820135
#>           0.2556438
#> result.10 0.2512543
#>           0.2161284
#> result.13       Inf
#>                 Inf
#>                 Inf
#>           0.7055311
#> 
#> Slot "ParetoFront":
#> [[1]]
#>                  [,1]      [,2]
#> result.1  0.029072714  59.57509
#> result.2  0.026890329 100.89280
#> result.3  0.003859345  76.48396
#> result.4  0.027466504  20.62048
#> result.5  0.019864753  47.13439
#> result.6  0.025348479  27.85536
#> result.7  0.108880468  14.49866
#> result.8  0.006243922 195.69724
#> result.9  0.050569465 185.49095
#> result.10 0.064984472 160.66696
#> result.11 0.050805401  35.34028
#> result.12 0.004287988 119.05332
#> result.13 0.051706671  48.16061
#> result.14 0.021283369 224.89984
#> 
#> [[2]]
#>                                
#> result.7  0.108880468  14.49866
#> result.14 0.021283369 224.89984
#> result.9  0.050569465 185.49095
#> result.13 0.092964459  76.69576
#> result.8  0.078191962  87.59060
#> result.9  0.065599078 151.89572
#> result.10 0.064984472 160.66696
#> result.2  0.073458271  65.89272
#> result.14 0.006159749 198.78661
#> result.15 0.063970858 115.59789
#> result.6  0.024414391 181.19219
#> result.8  0.006243922 195.69724
#> result.12 0.004287988 119.05332
#> result.13 0.051706671  48.16061
#> result.1  0.046506341 100.24508
#> 
#> [[3]]
#>                                
#> result.7  0.108880468  14.49866
#>           0.009397840 294.48435
#>           0.028894640 237.78186
#> result.13 0.092964459  76.69576
#> result.8  0.078191962  87.59060
#>           0.039515369 192.15347
#> result.9  0.065599078 151.89572
#> result.9  0.050569465 185.49095
#> result.10 0.064984472 160.66696
#>           0.002088606 246.68491
#>           0.075645449  59.48730
#>           0.051347347 138.06628
#> result.6  0.024414391 181.19219
#> result.15 0.063970858 115.59789
#>           0.023759291 230.56073
#> 
#> [[4]]
#>                                
#> result.7  0.108880468  14.49866
#>           0.009397840 294.48435
#>           0.028894640 237.78186
#>           0.085541004 132.75042
#>           0.108006062  61.41323
#>           0.039515369 192.15347
#> result.9  0.050569465 185.49095
#> result.9  0.065599078 151.89572
#>           0.095564224  96.13645
#> result.10 0.064984472 160.66696
#>           0.095167382 102.11118
#> result.13 0.092964459  76.69576
#>           0.002088606 246.68491
#>           0.051347347 138.06628
#> result.6  0.024414391 181.19219
#> 
#> [[5]]
#>                                 
#> result.7   0.108880468  14.49866
#>            0.009397840 294.48435
#>            0.028894640 237.78186
#>            0.085541004 132.75042
#>            0.108006062  61.41323
#>            0.039515369 192.15347
#> result.9   0.050569465 185.49095
#> result.9   0.065599078 151.89572
#>            0.095564224  96.13645
#> result.10  0.064984472 160.66696
#>            0.095167382 102.11118
#> result.13  0.092964459  76.69576
#>           -0.001285935 250.12797
#>           -0.001285935 250.12797
#>            0.051347347 138.06628
#> 
#>