A default call_back function that does nothing.

A default call_back function that does nothing.

callback_default(
  userdir = "",
  generation,
  pop_pool,
  pareto,
  prob_matrix,
  current_time
)

Arguments

userdir
the default directory used by galgo() to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with default_callback assigned to all the hook-points

galgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
start_galgo_callback = callback_default, # When Galgo is about to start.
end_galgo_callback = callback_default,   # When Galgo is about to finish. 
start_gen_callback = callback_default,   # At the beginning of each generation/iteration.
end_gen_callback = callback_default,     # At the end of each generation/iteration.
)

#> Using CPU for computing pearson distance

#> Generation 1 Non-dominated solutions:

#>            k                       rnkIndex     CrowD
#> result.1   4  0.11278306  23.56196        1       Inf
#> result.2   8  0.01141385 222.57181        1 0.8254984
#> result.4   4  0.08268501  74.01163        1 0.5426797
#> result.7  10 -0.01179783 259.92567        1       Inf
#> result.10  2  0.09501885  28.27588        1 0.4390350
#> result.11  4  0.06875280 113.06815        1 0.3898698
#> result.13  9  0.02673983 128.03740        1 0.7636250
#> result.14  6  0.05433006 115.47383        1 0.3958183

#> Generation 2 Non-dominated solutions:

#>            k                       rnkIndex     CrowD
#> result.1   4 0.112783057  23.56196        1       Inf
#> result.12 10 0.000283509 271.70962        1       Inf
#> result.2  10 0.019584412 167.21702        1 0.6215900
#> result.4   4 0.082685014  74.01163        1 0.5938833
#> result.2   8 0.011413854 222.57181        1 0.5458423
#> result.11  2 0.059499806 159.59322        1 0.4666480
#> result.10  2 0.095018850  28.27588        1 0.4303309
#> result.1   2 0.080086051 154.98218        1 0.3885443
#> result.15  4 0.072459559 155.93052        1 0.1824943

#> Generation 3 Non-dominated solutions:

#>            k                       rnkIndex     CrowD
#> result.1   4 0.112783057  23.56196        1       Inf
#> result.12 10 0.000283509 271.70962        1       Inf
#>            3 0.101115208 128.56398        1 0.7584870
#>            7 0.049846175 195.49971        1 0.6881006
#> result.2   8 0.011413854 222.57181        1 0.6854810
#>            6 0.073630973 172.57983        1 0.3956618
#> result.1   2 0.080086051 154.98218        1 0.3867465

#> Generation 4 Non-dominated solutions:

#>  k                      rnkIndex     CrowD
#>  3 0.13031964  44.33566        1       Inf
#>  6 0.02370174 338.29351        1       Inf
#>  7 0.04257734 218.49488        1 0.6292932
#>  3 0.10309314 145.49059        1 0.5280575
#>  3 0.09019366 195.27539        1 0.5096224
#>  7 0.04984617 195.49971        1 0.4074634
#>  3 0.10018432 145.66422        1 0.2460297

#> Generation 5 Non-dominated solutions:

#>  k                      rnkIndex     CrowD
#>  3 0.13031964  44.33566        1       Inf
#>  6 0.02370174 338.29351        1       Inf
#>  7 0.04257734 218.49488        1 0.6366426
#>  3 0.11169441 109.19127        1 0.5122546
#>  3 0.09019366 195.27539        1 0.5121874
#>  7 0.04984617 195.49971        1 0.4086585
#>  3 0.10018432 145.66422        1 0.2485920
#>  3 0.10309314 145.49059        1 0.1966151

#> NULL