Print minimal information to the user about galgo execution.

The main idea behind this callback function is to provide some feedback to the user about galgo execution. No other relevant information is shown

callback_no_report(
  userdir = "",
  generation,
  pop_pool,
  pareto,
  prob_matrix,
  current_time
)

Arguments

userdir
the default directory used by `galgo()` to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with no_report_callback assigned to the report_callback
# hook-point
galgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
report_callback = callback_no_report
)

#> Using CPU for computing pearson distance

#> ....*

#> An object of class "galgo.Obj"
#> Slot "Solutions":
#>          218060_s_at 208722_s_at 216283_s_at 212243_at 215077_at 221497_x_at
#>                    1           0           0         0         0           0
#>                    1           0           0         0         0           0
#>                    1           0           0         0         0           0
#>                    1           0           1         0         0           0
#>                    1           0           0         0         0           1
#>                    1           0           1         0         0           0
#> result.3           1           0           1         0         0           0
#>                    1           0           0         0         0           0
#> result.1           1           0           0         0         0           0
#>                    1           0           0         1         0           1
#>                    1           0           0         0         0           0
#>                    1           0           0         1         0           0
#> result.3           1           0           1         0         0           0
#>                    1           0           1         0         0           1
#>                    1           0           1         0         0           0
#>          201067_at 219632_s_at 219191_s_at 210579_s_at 216257_at 215707_s_at
#>                  0           0           0           0         0           0
#>                  0           0           0           0         1           1
#>                  1           0           0           0         1           0
#>                  0           0           0           1         0           0
#>                  0           0           0           1         1           0
#>                  0           0           0           0         0           0
#> result.3         0           0           0           0         0           0
#>                  1           0           0           0         1           0
#> result.1         1           0           0           0         1           0
#>                  0           0           0           1         1           0
#>                  0           0           0           1         1           0
#>                  0           0           0           1         1           0
#> result.3         0           0           0           0         0           0
#>                  0           0           0           1         0           0
#>                  0           0           0           0         0           0
#>          220678_at 201571_s_at 208245_at 222025_s_at 217904_s_at 219790_s_at
#>                  0           0         0           0           0           1
#>                  1           0         0           0           0           1
#>                  1           0         0           0           0           1
#>                  1           0         0           1           0           1
#>                  0           0         1           1           0           0
#>                  0           0         0           0           0           0
#> result.3         0           0         0           0           0           0
#>                  1           0         0           0           0           1
#> result.1         1           0         0           0           0           0
#>                  1           0         1           1           0           0
#>                  0           1         1           0           0           0
#>                  0           1         1           0           0           0
#> result.3         0           0         0           0           0           0
#>                  0           0         0           0           0           0
#>                  0           0         0           0           0           0
#>          211676_s_at 215801_at 215308_at 213261_at 201690_s_at 218703_at
#>                    0         0         1         1           0         0
#>                    1         0         0         0           0         1
#>                    0         0         1         0           0         1
#>                    0         0         1         0           0         1
#>                    0         0         1         0           0         0
#>                    1         0         1         1           0         0
#> result.3           1         0         1         1           0         0
#>                    0         0         1         0           0         1
#> result.1           1         0         1         0           0         1
#>                    0         0         0         0           0         0
#>                    1         0         0         0           1         0
#>                    0         0         0         0           1         0
#> result.3           1         0         1         1           0         0
#>                    1         0         0         1           0         0
#>                    1         0         1         1           0         0
#>          209968_s_at 212989_at 203602_s_at 204840_s_at 204720_s_at 213725_x_at
#>                    0         0           0           0           0           0
#>                    0         0           1           0           1           0
#>                    0         0           0           0           1           0
#>                    0         0           1           0           0           0
#>                    0         0           1           1           1           0
#>                    0         0           0           0           0           0
#> result.3           0         0           0           0           0           0
#>                    0         0           0           0           1           0
#> result.1           0         0           1           1           0           0
#>                    0         0           1           1           1           0
#>                    0         1           0           0           0           0
#>                    0         1           0           0           0           0
#> result.3           0         0           0           0           0           0
#>                    0         0           0           0           0           0
#>                    0         0           1           0           0           0
#>          213215_at 202173_s_at 201312_s_at 214289_at 212629_s_at 216933_x_at
#>                  0           0           0         0           0           0
#>                  0           1           0         0           0           1
#>                  1           1           0         1           0           0
#>                  0           0           0         1           0           0
#>                  0           1           0         1           0           1
#>                  0           0           0         0           0           0
#> result.3         0           0           0         0           0           0
#>                  1           1           0         1           0           0
#> result.1         1           1           0         1           0           1
#>                  1           1           0         1           0           1
#>                  0           0           0         0           0           0
#>                  0           0           0         0           0           0
#> result.3         0           0           0         0           0           0
#>                  0           0           0         0           0           0
#>                  0           0           0         0           0           0
#>          221663_x_at 218214_at 201518_at 212841_s_at 208000_at 209131_s_at
#>                    0         0         0           0         0           0
#>                    0         0         0           0         0           0
#>                    0         0         1           0         0           0
#>                    1         1         1           0         0           1
#>                    0         0         1           0         0           0
#>                    0         0         0           0         0           0
#> result.3           0         0         0           0         0           0
#>                    0         0         1           0         0           0
#> result.1           0         0         0           0         0           0
#>                    0         0         1           0         0           0
#>                    0         0         0           0         0           0
#>                    0         0         0           0         0           0
#> result.3           0         0         0           0         0           0
#>                    0         0         1           0         0           0
#>                    0         0         0           0         0           0
#>          217494_s_at 221588_x_at 210428_s_at 205037_at 204947_at 219682_s_at
#>                    0           0           0         0         0           0
#>                    1           1           1         0         0           1
#>                    1           0           1         0         0           1
#>                    0           1           0         0         0           0
#>                    0           0           1         0         0           0
#>                    0           0           0         0         0           0
#> result.3           0           0           0         0         0           0
#>                    1           0           0         0         0           1
#> result.1           1           0           1         0         0           1
#>                    0           0           1         0         0           0
#>                    1           1           1         0         0           1
#>                    1           1           1         0         0           1
#> result.3           0           0           0         0         0           0
#>                    0           1           0         0         0           0
#>                    0           0           0         0         0           1
#>          203468_at 216821_at 211670_x_at 214239_x_at 219372_at 209483_s_at
#>                  0         0           0           1         1           0
#>                  1         0           0           0         0           0
#>                  0         0           0           0         0           1
#>                  0         0           0           0         1           0
#>                  0         0           0           1         0           0
#>                  0         0           0           0         1           0
#> result.3         0         0           0           0         1           0
#>                  0         0           0           0         1           1
#> result.1         0         0           0           1         0           1
#>                  1         0           0           1         0           0
#>                  1         0           0           0         0           0
#>                  1         0           0           0         0           0
#> result.3         0         0           0           1         1           0
#>                  1         0           0           0         1           0
#>                  0         0           0           0         1           0
#>          219749_at 221498_at 221965_at 203688_at 201596_x_at 220517_at
#>                  0         0         0         0           0         0
#>                  1         0         1         0           1         0
#>                  1         0         1         0           1         0
#>                  0         0         0         0           0         0
#>                  0         0         0         0           1         0
#>                  0         0         0         0           0         0
#> result.3         0         0         0         0           0         0
#>                  0         0         0         0           1         0
#> result.1         1         0         1         0           1         0
#>                  1         0         0         0           1         0
#>                  0         0         0         0           0         0
#>                  1         0         0         0           0         0
#> result.3         0         0         0         0           0         0
#>                  0         0         0         0           0         0
#>                  0         0         0         0           0         0
#>          213978_at 217168_s_at 203929_s_at 200778_s_at 204818_at 205456_at
#>                  0           0           0           0         0         0
#>                  0           1           0           1         0         0
#>                  0           0           0           1         0         1
#>                  0           0           0           0         0         0
#>                  0           0           0           0         1         0
#>                  0           0           0           0         0         0
#> result.3         0           0           0           0         0         0
#>                  0           0           0           1         0         0
#> result.1         0           0           0           1         0         1
#>                  0           0           0           1         1         0
#>                  0           0           0           1         0         1
#>                  0           0           1           1         0         0
#> result.3         0           0           0           0         0         0
#>                  0           0           0           0         0         0
#>                  0           0           0           0         0         0
#>          209878_s_at 221946_at 217057_s_at 218622_at 221194_s_at 206527_at
#>                    0         1           0         0           0         0
#>                    0         1           0         0           1         0
#>                    0         1           1         0           1         0
#>                    0         0           0         0           0         0
#>                    0         1           1         0           1         0
#>                    0         0           0         0           0         0
#> result.3           0         0           0         0           0         0
#>                    0         1           1         0           0         0
#> result.1           0         0           1         0           1         0
#>                    0         1           1         0           1         0
#>                    0         0           0         0           1         0
#>                    0         0           0         0           1         0
#> result.3           0         0           0         0           0         0
#>                    0         0           0         0           0         0
#>                    0         0           0         0           0         0
#>          212907_at 205642_at 212612_at 210244_at 201816_s_at 210926_at
#>                  0         0         0         0           0         0
#>                  1         1         0         0           1         0
#>                  1         0         0         0           0         0
#>                  0         1         0         0           0         0
#>                  1         0         0         0           1         0
#>                  0         0         0         0           0         0
#> result.3         0         0         0         0           0         0
#>                  1         0         0         0           0         0
#> result.1         1         0         0         0           0         0
#>                  1         0         0         0           1         0
#>                  0         1         0         0           1         0
#>                  0         1         0         0           1         0
#> result.3         0         0         0         0           0         0
#>                  0         1         0         0           0         0
#>                  0         0         0         0           0         0
#>          219238_at 208063_s_at 216043_x_at 219716_at 207884_at 219698_s_at
#>                  0           1           0         0         1           0
#>                  1           1           0         0         0           0
#>                  1           1           1         0         1           1
#>                  0           1           1         1         1           0
#>                  1           0           0         0         0           0
#>                  0           1           0         0         1           0
#> result.3         0           1           0         0         1           0
#>                  1           1           1         0         1           1
#> result.1         1           0           0         0         0           1
#>                  1           0           0         0         0           0
#>                  0           1           0         0         1           0
#>                  0           1           0         0         1           0
#> result.3         0           1           0         0         1           0
#>                  0           1           0         1         1           0
#>                  0           1           0         0         1           0
#>          202795_x_at 203018_s_at 213849_s_at 203149_at 208524_at 218355_at
#>                    1           0           0         0         0         0
#>                    0           0           0         0         1         0
#>                    1           0           0         0         0         1
#>                    0           0           0         1         0         0
#>                    1           1           0         0         0         1
#>                    1           0           0         0         0         0
#> result.3           1           0           0         0         0         0
#>                    1           0           0         0         0         0
#> result.1           0           1           0         0         1         0
#>                    1           1           0         0         0         1
#>                    1           0           0         0         1         0
#>                    1           0           0         0         1         0
#> result.3           1           0           0         0         0         0
#>                    1           0           1         0         0         0
#>                    1           0           0         0         0         1
#>          218614_at 221831_at 204948_s_at 204370_at 206118_at 204796_at
#>                  0         0           0         0         0         0
#>                  0         0           1         1         0         0
#>                  1         1           0         0         1         1
#>                  1         0           1         0         0         1
#>                  1         0           0         0         1         1
#>                  0         0           1         0         0         0
#> result.3         0         0           1         0         0         0
#>                  1         1           0         0         0         1
#> result.1         0         1           0         0         1         1
#>                  1         1           0         0         1         1
#>                  0         0           1         0         0         0
#>                  0         0           1         0         0         0
#> result.3         0         0           1         0         0         0
#>                  1         0           1         0         0         0
#>                  0         0           1         0         0         0
#>          221661_at 201843_s_at 201315_x_at 221820_s_at  k             
#>                  0           0           0           0  3  0.180863100
#>                  0           0           0           0 10  0.013712525
#>                  0           0           1           0  6  0.053904706
#>                  0           1           0           1  7  0.037525749
#>                  0           0           0           0  2  0.133102675
#>                  0           0           0           0  7  0.149524624
#> result.3         0           0           0           0  5  0.140004158
#>                  0           0           1           0  2  0.133104266
#> result.1         0           0           1           0  8  0.003613777
#>                  0           0           0           0  2  0.129878508
#>                  0           0           0           0 10  0.050257673
#>                  0           0           0           0 10  0.034669125
#> result.3         0           0           0           0  4  0.102467273
#>                  0           0           0           1  9 -0.011747623
#>                  0           0           0           0  2  0.118466814
#>                    rnkIndex     CrowD
#>           61.10022        1       Inf
#>          306.62388        1       Inf
#>          181.37060        1 0.7455274
#>          231.61563        1 0.6379807
#>          158.87844        1 0.6088158
#>           76.93364        1 0.3433948
#> result.3  99.39671        1 0.2455870
#>          123.71226        1 0.2397070
#> result.1 240.59611        2       Inf
#>          120.96666        2       Inf
#>          154.09082        2 0.6340347
#>          231.26857        2 0.5386667
#> result.3 148.98236        2 0.5269112
#>          220.97949        3       Inf
#>           84.63171        3       Inf
#> 
#> Slot "ParetoFront":
#> [[1]]
#>                  [,1]       [,2]
#> result.1  0.038657943   6.711632
#> result.2  0.037835876  23.991271
#> result.3  0.140004158  99.396712
#> result.4  0.100461147 108.234282
#> result.5  0.018680071 185.785494
#> result.6  0.052904348  64.615036
#> result.7  0.025392820 105.377292
#> result.8  0.051285055  53.806622
#> result.9  0.009871397 156.328509
#> result.10 0.040181116   2.338172
#> result.11 0.084529534   8.803596
#> result.12 0.021019972  69.839643
#> result.13 0.058462805   9.614301
#> result.14 0.047048556 145.025797
#> 
#> [[2]]
#>                                
#> result.3  0.140004158  99.39671
#> result.1  0.003613777 240.59611
#> result.3  0.102467273 148.98236
#> result.14 0.018829436 165.28191
#> result.5  0.018680071 185.78549
#> result.4  0.100461147 108.23428
#> result.9  0.009871397 156.32851
#> result.14 0.047048556 145.02580
#> result.6  0.099295289 103.03712
#> result.9  0.007796874 132.06635
#> result.10 0.036823820 121.35812
#> result.7  0.025392820 105.37729
#> result.15 0.092886065  23.96932
#> result.6  0.052904348  64.61504
#> result.4  0.076169670  29.46784
#> 
#> [[3]]
#>                                 
#> result.1  0.0036137765 240.59611
#>           0.1808630998  61.10022
#> result.3  0.1024672733 148.98236
#>           0.0539047060 181.37060
#> result.5  0.0186800708 185.78549
#>           0.1298785077 120.96666
#>           0.1495246238  76.93364
#> result.3  0.1400041583  99.39671
#> result.4  0.1004611466 108.23428
#>           0.0006036237 189.92552
#>           0.0614078746 130.56122
#> result.14 0.0188294363 165.28191
#> result.14 0.0470485558 145.02580
#> result.9  0.0098713967 156.32851
#> result.6  0.0992952895 103.03712
#> 
#> [[4]]
#>                                
#>           0.180863100  61.10022
#>           0.013712525 306.62388
#>           0.053904706 181.37060
#>           0.133102675 158.87844
#>           0.037525749 231.61563
#> result.3  0.140004158  99.39671
#>           0.149524624  76.93364
#> result.1  0.003613777 240.59611
#>           0.129878508 120.96666
#>           0.050257673 154.09082
#>           0.034669125 231.26857
#> result.3  0.102467273 148.98236
#>          -0.011747623 220.97949
#>           0.118466814  84.63171
#> result.4  0.100461147 108.23428
#> 
#> [[5]]
#>                                
#>           0.180863100  61.10022
#>           0.013712525 306.62388
#>           0.053904706 181.37060
#>           0.037525749 231.61563
#>           0.133102675 158.87844
#>           0.149524624  76.93364
#> result.3  0.140004158  99.39671
#>           0.133104266 123.71226
#> result.1  0.003613777 240.59611
#>           0.129878508 120.96666
#>           0.050257673 154.09082
#>           0.034669125 231.26857
#> result.3  0.102467273 148.98236
#>          -0.011747623 220.97949
#>           0.118466814  84.63171
#> 
#>