Classify samples from multiple centroids

Classify samples from multiple centroids

classify_multiple(prob_matrix, centroid._list, distancetype = "pearson")

Arguments

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

centroid._list
alist with the centroid matrix for each of the signatures to evaluate, where each column represents the prototypic centroid of a subtype and each row the constituents features of the solution signature. The output of create_centroids can be used.

distancetype
a character that can be either 'pearson' (default), 'spearman' or 'kendall'.

Value

Returns a data.frame with the classes assigned to each sample in each signature, were samples are a rows and signatures in columns

Examples

if (FALSE) {
# Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time = TCGA_clinic$time, event = TCGA_clinic$status)

# Run galgo
output <- galgoR::galgo(generations = 10, population = 30, prob_matrix = TCGA_expr, OS = OS)
RESULTS <- non_dominated_summary(
  output = output, OS = OS,
  prob_matrix = TCGA_expr,
  distancetype = "pearson",
  usegpu = FALSE
)
CentroidsList <- create_centroids(output, RESULTS$solution, train.set = TCGA_expr)
TCGA_classes <- classify_multiple(prob_matrix = TCGA_expr, centroid._list = CentroidsList)
}