Wrapper function to perform partition around medioids (PAM) for GalgoR

In galgoR, the partition around medioids (PAM) algorithm is the default clustering process used under the evolutionary process.

cluster_algorithm(c, k)

Arguments

c
a dissimilarity matrix object of type 'dist'

k
positive integer specifying the number of clusters, less than the number of observations

Value

Returns a 'list' with the value '$cluster' which contains the cluster assignment of each of the samples evaluated

Details

The function runs the pam function of the 'cluster' package with options cluster.only =TRUE, diss = TRUE, do.swap=TRUE, keep.diss=FALSE, keep.data = FALSE, pamonce= 2

References

Reynolds, A., Richards, G., de la Iglesia, B. and Rayward-Smith, V. (1992) Clustering rules: A comparison of partitioning and hierarchical clustering algorithms; Journal of Mathematical Modelling and Algorithms 5, 475--504. 10.1007/s10852-005-9022-1.

Erich Schubert and Peter J. Rousseeuw (2019) Faster k-Medoids Clustering: Improving the PAM, CLARA, and CLARANS Algorithms; Preprint, (https://arxiv.org/abs/1810.05691).

Examples

if (FALSE) {
rna_luad <- use_rna_luad()
prm <- rna_luad$TCGA$expression_matrix
Dist <- calculate_distance_pearson_cpu(prm)
k <- 4
Pam <- cluster_algorithm(Dist, k)
table(Pam$cluster)
}