GalgoR main function

galgo accepts an expression matrix and a survival object to find robust gene expression signatures related to a given outcome

galgo(
  population = 30,
  generations = 2,
  nCV = 5,
  usegpu = FALSE,
  distancetype = "pearson",
  TournamentSize = 2,
  period = 1825,
  OS,
  prob_matrix,
  res_dir = "",
  start_galgo_callback = default_callback,
  end_galgo_callback = base_return_pop_callback,
  report_callback = base_report_callback,
  start_gen_callback = default_callback,
  end_gen_callback = default_callback,
  verbose = 2
)

Arguments

population
a number indicating the number of solutions in the population of solutions that will be evolved

generations
a number indicating the number of iterations of the galgo algorithm

nCV
number of cross-validation sets

usegpu
logical default to FALSE, set to TRUE if you wish to use gpu computing (gpuR package must be properly installed and loaded)

distancetype
character, it can be 'pearson' (centered pearson), 'uncentered' (uncentered pearson), 'spearman' or 'euclidean'

TournamentSize
a number indicating the size of the tournaments for the selection procedure

period
a number indicating the outcome period to evaluate the RMST

OS
a survival object (see Surv function from the survival package)

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

res_dir
a character string indicating where to save the intermediate and final output of the algorithm

start_galgo_callback
optional callback function for the start of the galgo execution

end_galgo_callback
optional callback function for the end of the galgo execution

report_callback
optional callback function

start_gen_callback
optional callback function for the beginning of the run

end_gen_callback
optional callback function for the end of the run

verbose
select the level of information printed during galgo execution

Value

an object of type 'galgo.Obj' that corresponds to a list with the elements $Solutions and $ParetoFront. $Solutions is a \(l x (n + 5)\) matrix where \(n\) is the number of features evaluated and \(l\) is the number of solutions obtained. The submatrix \(l x n\) is a binary matrix where each row represents the chromosome of an evolved solution from the solution population, where each feature can be present (1) or absent (0) in the solution. Column \(n +1\) represent the \(k\) number of clusters for each solutions. Column \(n+2\) to \(n+5\) shows the SC Fitness and Survival Fitness values, the solution rank, and the crowding distance of the solution in the final pareto front respectively. For easier interpretation of the 'galgo.Obj', the output can be reshaped using the to_list and to_dataframe functions

Examples

# load example dataset
library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)

# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo
output <- galgoR::galgo(generations = 5, population = 15, prob_matrix = expression, OS = OS)

#> Using CPU for computing pearson distance

#> Generation 1 Non-dominated solutions:

#>            k                      rnkIndex    CrowD
#> result.1   3 0.09073725 128.82984        1 1.023008
#> result.3   8 0.02212960 213.55445        1 0.389267
#> result.9   4 0.14558053  24.92632        1      Inf
#> result.11 10 0.02043814 245.97382        1      Inf
#> result.12  4 0.05796947 210.71596        1 0.573795
#> result.13  3 0.08510381 172.61869        1 0.562403

#> Generation 2 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.9   4 0.14558053  24.92632        1       Inf
#> result.6   8 0.01210128 257.02469        1       Inf
#> result.8   2 0.10883566 172.95062        1 1.2184884
#> result.3   4 0.06274179 204.31701        1 0.4664650
#> result.3   8 0.02212960 213.55445        1 0.3729520
#> result.12  4 0.05796947 210.71596        1 0.2910301
#> result.11 10 0.02043814 245.97382        1 0.2321734

#> Generation 3 Non-dominated solutions:

#>          k                      rnkIndex    CrowD
#> result.9 4 0.14558053  24.92632        1      Inf
#>          4 0.07021006 319.43006        1      Inf
#> result.8 2 0.10883566 172.95062        1 1.395612

#> Generation 4 Non-dominated solutions:

#>          k                      rnkIndex    CrowD
#> result.9 4 0.14558053  24.92632        1      Inf
#>          4 0.07021006 319.43006        1      Inf
#> result.8 2 0.10883566 172.95062        1 1.427828

#> Generation 5 Non-dominated solutions:

#>          k                      rnkIndex     CrowD
#> result.9 4 0.14558053  24.92632        1       Inf
#>          4 0.07021006 319.43006        1       Inf
#>          4 0.13090785 180.21854        1 1.1096426
#>          3 0.09675105 274.67703        1 0.8431102

outputDF <- to_dataframe(output)
outputList <- to_list(output)