Print minimal information to the user about galgo execution.

The main idea behind this callback function is to provide some feedback to the user about galgo execution. No other relevant information is shown

no_report_callback(
  userdir = "",
  generation,
  pop_pool,
  pareto,
  prob_matrix,
  current_time
)

Arguments

userdir
the default directory used by `galgo()` to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with no_report_callback assigned to the report_callback
# hook-point
galgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
report_callback = no_report_callback
)

#> Using CPU for computing pearson distance

#> ....*

#> An object of class "galgo.Obj"
#> Slot "Solutions":
#>           203506_s_at 209660_at 38043_at 216018_at 206434_at 213141_at
#> result.8            0         1        1         1         1         1
#>                     0         0        0         0         1         0
#>                     1         1        0         0         0         1
#> result.14           1         0        1         1         0         1
#> result.9            1         1        1         1         1         1
#> result.7            0         0        0         0         0         0
#>                     1         1        1         1         1         1
#> result.8            0         1        0         0         1         1
#> result.1            0         1        0         1         0         1
#>                     1         1        1         1         1         1
#>                     0         1        1         1         1         1
#> result.12           1         1        1         1         1         1
#>                     1         1        1         1         1         1
#> result.10           1         1        1         1         1         1
#> result.1            0         1        0         0         0         0
#>           201836_s_at 222011_s_at 219114_at 215902_at 210233_at 214820_at
#> result.8            1           1         1         1         1         1
#>                     0           0         1         0         0         0
#>                     0           0         1         0         0         0
#> result.14           1           1         1         0         1         1
#> result.9            0           1         1         1         1         0
#> result.7            0           0         1         0         0         0
#>                     1           1         1         1         1         1
#> result.8            0           1         1         0         1         1
#> result.1            1           1         1         1         0         0
#>                     1           1         1         1         1         1
#>                     1           1         1         1         1         1
#> result.12           1           1         1         0         1         1
#>                     1           1         1         1         1         1
#> result.10           1           1         1         1         1         1
#> result.1            1           1         1         0         0         0
#>           200679_x_at 219600_s_at 206317_s_at 219014_at 204957_at 214245_at
#> result.8            1           1           1         1         1         1
#>                     0           0           1         0         0         1
#>                     0           0           0         0         0         0
#> result.14           1           0           1         1         1         1
#> result.9            1           1           0         1         1         1
#> result.7            0           0           1         0         0         1
#>                     1           1           0         1         1         1
#> result.8            0           0           1         0         0         1
#> result.1            0           0           0         0         1         1
#>                     1           0           1         1         1         1
#>                     1           1           1         1         1         1
#> result.12           0           0           0         1         1         0
#>                     0           0           0         1         1         0
#> result.10           1           0           0         1         1         1
#> result.1            0           0           0         0         1         0
#>           215755_at 204335_at 200639_s_at 211026_s_at 218461_at 202524_s_at
#> result.8          1         1           1           1         1           1
#>                   0         1           0           0         0           0
#>                   0         0           1           0         0           1
#> result.14         0         1           1           1         1           1
#> result.9          1         1           1           1         0           0
#> result.7          0         1           0           0         0           0
#>                   0         1           1           1         0           1
#> result.8          1         1           0           1         0           0
#> result.1          0         0           1           1         0           0
#>                   0         1           1           1         0           1
#>                   1         1           1           1         1           1
#> result.12         0         1           1           1         0           1
#>                   0         1           1           1         0           1
#> result.10         0         1           1           1         0           1
#> result.1          0         0           0           1         0           0
#>           204433_s_at 215799_at 206586_at 202068_s_at 205177_at 203355_s_at
#> result.8            1         1         1           1         1           0
#>                     0         0         0           0         0           0
#>                     0         0         1           1         1           0
#> result.14           1         0         0           1         1           1
#> result.9            0         0         1           1         1           1
#> result.7            0         0         0           0         0           0
#>                     1         0         0           0         0           1
#> result.8            1         0         1           0         0           0
#> result.1            0         0         0           0         0           1
#>                     1         1         1           0         1           0
#>                     1         1         1           1         0           1
#> result.12           1         0         0           0         0           1
#>                     1         1         0           0         0           1
#> result.10           1         1         1           0         0           1
#> result.1            0         0         0           0         0           0
#>           210831_s_at 217779_s_at 204654_s_at 220190_s_at 203818_s_at
#> result.8            1           1           1           1           1
#>                     0           1           0           0           0
#>                     0           1           0           0           0
#> result.14           0           1           1           0           1
#> result.9            0           1           1           1           1
#> result.7            0           1           0           0           0
#>                     1           1           1           1           1
#> result.8            0           1           1           0           1
#> result.1            0           1           0           1           1
#>                     1           1           1           1           1
#>                     1           1           1           1           1
#> result.12           0           1           1           0           1
#>                     0           1           1           0           1
#> result.10           1           1           1           1           1
#> result.1            0           1           0           1           1
#>           210961_s_at 203234_at 222247_at 210576_at 210781_x_at 213855_s_at
#> result.8            1         1         1         1           1           1
#>                     1         0         0         0           0           0
#>                     0         0         0         0           0           0
#> result.14           1         1         1         0           1           0
#> result.9            1         0         1         1           1           1
#> result.7            0         0         0         0           0           0
#>                     0         1         1         1           1           1
#> result.8            0         0         1         0           0           0
#> result.1            0         1         1         1           1           1
#>                     1         1         1         0           1           0
#>                     1         1         1         1           1           1
#> result.12           0         1         1         0           1           0
#>                     1         1         0         0           1           0
#> result.10           1         1         1         0           1           0
#> result.1            0         1         0         0           1           1
#>           218324_s_at 201058_s_at 218667_at 211775_x_at 208406_s_at 204100_at
#> result.8            1           1         1           1           1         1
#>                     0           0         0           0           0         0
#>                     1           0         1           0           1         0
#> result.14           1           0         1           0           1         1
#> result.9            1           0         1           0           1         1
#> result.7            0           0         0           0           0         0
#>                     1           0         1           0           1         1
#> result.8            0           1         1           0           0         1
#> result.1            0           0         1           0           0         1
#>                     1           1         1           0           1         1
#>                     1           1         1           1           1         1
#> result.12           1           0         1           0           1         1
#>                     1           1         1           0           1         1
#> result.10           1           0         1           0           1         1
#> result.1            0           0         0           0           0         0
#>           214201_x_at 207916_at 200752_s_at 220024_s_at 213346_at 217931_at
#> result.8            1         1           1           1         1         1
#>                     0         0           1           0         1         1
#>                     0         0           0           0         0         0
#> result.14           1         1           0           1         0         0
#> result.9            1         0           0           1         1         1
#> result.7            0         0           1           0         1         1
#>                     1         1           1           1         0         0
#> result.8            1         0           1           1         1         1
#> result.1            0         0           0           1         1         0
#>                     1         0           0           1         0         0
#>                     1         1           1           1         0         1
#> result.12           1         1           0           1         0         0
#>                     1         0           0           1         0         0
#> result.10           1         0           0           1         0         0
#> result.1            0         1           0           0         1         0
#>           206947_at 216035_x_at 217665_at 213421_x_at 209903_s_at 202592_at
#> result.8          1           1         1           1           1         1
#>                   0           0         0           0           0         0
#>                   0           0         0           0           0         0
#> result.14         1           1         1           0           1         0
#> result.9          1           1         1           1           0         0
#> result.7          0           0         0           0           0         0
#>                   1           1         1           0           1         1
#> result.8          0           0         1           1           0         1
#> result.1          0           0         1           1           0         1
#>                   0           1         1           1           0         1
#>                   1           1         1           0           1         1
#> result.12         1           1         1           0           1         0
#>                   1           1         1           0           0         1
#> result.10         0           1         1           0           0         1
#> result.1          0           0         0           0           0         0
#>           202008_s_at 217977_at 207389_at 212443_at 202258_s_at 55065_at
#> result.8            1         1         1         1           0        0
#>                     0         0         0         0           0        0
#>                     0         0         0         0           1        0
#> result.14           1         0         1         0           1        0
#> result.9            1         1         1         0           1        1
#> result.7            0         0         0         0           0        0
#>                     1         0         1         1           1        1
#> result.8            1         0         1         1           0        0
#> result.1            1         1         1         0           1        1
#>                     1         1         1         0           1        0
#>                     1         0         1         1           1        0
#> result.12           1         0         1         0           1        1
#>                     1         0         1         0           1        1
#> result.10           1         0         1         0           1        0
#> result.1            1         0         0         0           1        1
#>           205702_at 219902_at 218560_s_at 203836_s_at 209561_at 211632_at
#> result.8          1         1           1           1         1         1
#>                   0         0           0           1         0         1
#>                   1         0           0           1         0         0
#> result.14         1         0           0           1         1         1
#> result.9          1         1           1           0         0         1
#> result.7          0         0           0           0         0         1
#>                   0         0           0           1         1         1
#> result.8          0         0           0           0         1         1
#> result.1          1         0           1           1         0         0
#>                   1         0           0           1         1         1
#>                   1         1           0           1         1         1
#> result.12         0         0           0           1         1         1
#>                   1         0           0           1         1         1
#> result.10         1         0           0           1         1         1
#> result.1          0         0           1           1         0         0
#>           201911_s_at 210794_s_at 210153_s_at 208266_at 218047_at 206716_at
#> result.8            1           1           1         1         1         1
#>                     0           1           0         0         0         1
#>                     0           0           1         1         0         0
#> result.14           1           1           1         1         1         1
#> result.9            1           1           0         1         1         0
#> result.7            0           1           0         0         0         1
#>                     1           1           1         1         1         1
#> result.8            0           1           0         0         0         1
#> result.1            0           1           1         0         0         0
#>                     1           1           0         1         1         1
#>                     1           1           0         1         1         1
#> result.12           1           1           1         0         1         1
#>                     1           1           0         0         1         1
#> result.10           1           1           0         1         1         1
#> result.1            0           0           0         0         0         0
#>           214962_s_at 209959_at 209367_at 212834_at 218831_s_at 211490_at
#> result.8            0         1         1         0           1         1
#>                     0         1         0         0           0         0
#>                     0         0         0         0           0         1
#> result.14           1         1         1         0           1         0
#> result.9            1         1         0         1           0         0
#> result.7            0         1         0         0           0         1
#>                     1         1         0         0           0         0
#> result.8            0         1         0         1           1         1
#> result.1            1         0         1         1           0         1
#>                     1         1         1         0           1         0
#>                     1         1         1         0           0         1
#> result.12           1         1         0         0           0         0
#>                     1         1         0         0           0         0
#> result.10           1         1         1         0           1         0
#> result.1            1         0         0         1           0         1
#>           218238_at 221676_s_at 218573_at 213875_x_at 206990_at 222153_at
#> result.8          1           1         1           1         1         1
#>                   0           1         0           0         0         0
#>                   0           0         0           0         0         0
#> result.14         0           1         1           1         0         0
#> result.9          0           0         1           1         0         0
#> result.7          0           0         0           0         0         0
#>                   0           1         1           0         0         0
#> result.8          1           1         1           1         0         0
#> result.1          1           0         0           0         1         0
#>                   1           1         1           1         1         0
#>                   1           1         1           1         1         0
#> result.12         0           1         1           0         0         0
#>                   0           1         1           0         1         0
#> result.10         0           1         1           1         1         0
#> result.1          1           0         0           0         0         0
#>           202494_at 201149_s_at 216224_s_at 201222_s_at 213189_at 211605_s_at
#> result.8          1           1           1           1         1           1
#>                   0           0           0           0         0           0
#>                   0           0           1           0         0           1
#> result.14         0           1           0           0         1           1
#> result.9          1           0           0           1         1           0
#> result.7          0           0           0           0         0           0
#>                   1           0           1           0         1           1
#> result.8          1           0           1           1         0           0
#> result.1          1           0           0           0         0           1
#>                   0           1           0           0         1           1
#>                   1           1           1           1         1           1
#> result.12         0           0           0           0         1           1
#>                   0           0           0           0         1           1
#> result.10         0           1           0           0         1           1
#> result.1          0           0           0           0         0           1
#>           216788_at 209669_s_at 210632_s_at 214719_at 208947_s_at  k
#> result.8          1           1           1         1           1 10
#>                   0           0           0         0           1  2
#>                   0           0           0         0           0  4
#> result.14         1           1           1         0           1  7
#> result.9          1           1           1         1           1  6
#> result.7          0           0           0         0           0  2
#>                   1           1           1         0           1 10
#> result.8          0           0           1         0           0  2
#> result.1          0           0           0         0           1  5
#>                   0           1           1         0           1  2
#>                   1           1           1         1           1 10
#> result.12         0           1           1         0           1  8
#>                   0           1           1         0           1  2
#> result.10         0           1           1         0           1  5
#> result.1          0           0           0         0           0  5
#>                                 rnkIndex     CrowD
#> result.8  -0.003626591 329.1283        1       Inf
#>            0.117963145 186.1210        1       Inf
#>            0.073046537 222.6191        1 0.8689906
#> result.14  0.029276277 265.3073        1 0.5919964
#> result.9   0.022268603 298.9930        1 0.4272188
#> result.7   0.109268162 102.6672        2       Inf
#>            0.010013861 221.8478        2       Inf
#> result.8   0.083748461 176.5009        2 0.6581919
#> result.1   0.023328683 191.8677        2 0.6249388
#>            0.097478327 148.1222        2 0.4054109
#>            0.013241080 217.7946        2 0.1855238
#> result.12  0.004524431 208.2488        3       Inf
#>            0.096284133  83.4890        3       Inf
#> result.10  0.061084953 131.8985        3 0.4137383
#> result.1   0.035675144 162.3784        3 0.4018746
#> 
#> Slot "ParetoFront":
#> [[1]]
#>                    [,1]       [,2]
#> result.1   0.0356751436 162.378445
#> result.2  -0.0002325493  25.084800
#> result.3  -0.0157067264 126.132170
#> result.4   0.0138296688  39.080717
#> result.5   0.0444034647 137.437453
#> result.6   0.0088833112  99.629015
#> result.7   0.1092681622 102.667236
#> result.8  -0.0036265908 329.128315
#> result.9   0.0222686031 298.993005
#> result.10  0.0079519723  18.104396
#> result.11  0.0445516886   5.563439
#> result.12  0.0252553505  26.437668
#> result.13  0.0429342871  88.308006
#> result.14  0.0613397992  86.145788
#> result.15  0.0311017835  82.984682
#> 
#> [[2]]
#>                                 
#> result.7   0.109268162 102.66724
#> result.8  -0.003626591 329.12832
#> result.8   0.083748461 176.50095
#> result.14  0.029276277 265.30734
#> result.9   0.022268603 298.99301
#> result.14  0.061339799  86.14579
#> result.12  0.004524431 208.24875
#> result.1   0.023328683 191.86774
#> result.1   0.035675144 162.37844
#> result.5   0.044403465 137.43745
#> result.10  0.061084953 131.89851
#> result.5   0.049949305  21.44234
#> result.15 -0.017504952 161.87653
#> result.3   0.019716226 116.98455
#> result.13  0.042934287  88.30801
#> 
#> [[3]]
#>                                 
#> result.7   0.109268162 102.66724
#> result.8  -0.003626591 329.12832
#>            0.073046537 222.61912
#> result.14  0.029276277 265.30734
#> result.9   0.022268603 298.99301
#>            0.097478327 148.12216
#> result.8   0.083748461 176.50095
#> result.12  0.004524431 208.24875
#>            0.096284133  83.48900
#> result.14  0.061339799  86.14579
#> result.1   0.023328683 191.86774
#> result.1   0.035675144 162.37844
#> result.5   0.044403465 137.43745
#> result.10  0.061084953 131.89851
#>           -0.009369553 182.34447
#> 
#> [[4]]
#>                                 
#> result.8  -0.003626591 329.12832
#>            0.117963145 186.12102
#>            0.073046537 222.61912
#> result.14  0.029276277 265.30734
#> result.9   0.022268603 298.99301
#> result.7   0.109268162 102.66724
#>            0.013241080 217.79465
#> result.8   0.083748461 176.50095
#> result.1   0.023328683 191.86774
#>            0.097478327 148.12216
#> result.12  0.004524431 208.24875
#>            0.096284133  83.48900
#> result.1   0.035675144 162.37844
#> result.14  0.061339799  86.14579
#> result.10  0.061084953 131.89851
#> 
#> [[5]]
#>                                
#> result.8  -0.003626591 329.1283
#>            0.117963145 186.1210
#>            0.073046537 222.6191
#> result.14  0.029276277 265.3073
#> result.9   0.022268603 298.9930
#> result.7   0.109268162 102.6672
#>            0.010013861 221.8478
#> result.8   0.083748461 176.5009
#> result.1   0.023328683 191.8677
#>            0.097478327 148.1222
#>            0.013241080 217.7946
#> result.12  0.004524431 208.2488
#>            0.096284133  83.4890
#> result.10  0.061084953 131.8985
#> result.1   0.035675144 162.3784
#> 
#>