Summary of the non dominated solutions

The function uses a 'galgo.Obj' as input an the training dataset to evaluate the non-dominated solutions found by GalgoR

non_dominated_summary(
  output,
  prob_matrix,
  OS,
  distancetype = "pearson",
  usegpu = FALSE
)

Arguments

output
An object of class galgo.Obj

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

OS
a survival object (see Surv function from the survival package)

distancetype
a character that can be either 'pearson', 'uncentered', 'spearman' or 'euclidean'

usegpu
logical TRUE or FALSE

Value

Returns a data.frame with 5 columns and a number of rows equals to the non-dominated solutions found by GalgoR. The first column has the name of the non-dominated solutions, the second the number of partitions found for each solution (k), the third, the number of genes, the fourth the mean silhouette coefficient of the solution and the last columns has the estimated C.Index for each one.

Examples

if (FALSE) {
# Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time = TCGA_clinic$time, event = TCGA_clinic$status)

# Run galgo
output <- galgoR::galgo(generations = 10, population = 30, prob_matrix = TCGA_expr, OS = OS)
non_dominated_summary(
  output = output, OS = OS,
  prob_matrix = TCGA_expr,
  distancetype = "pearson",
  usegpu = FALSE
)
}