Print minimal information to the user about galgo execution.

The main idea behind this callback function is to provide some feedback to the user about galgo execution. No other relevant information is shown

callback_no_report(userdir = "", generation, pop_pool, pareto,
  prob_matrix, current_time)

Arguments

userdir
the default directory used by `galgo()` to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with no_report_callback assigned to the report_callback
# hook-point
galgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
report_callback = callback_no_report
)

#> Using CPU for computing pearson distance

#> ....*

#> An object of class "galgo.Obj"
#> Slot "Solutions":
#>           208276_at 211240_x_at 201006_at 209307_at 215845_x_at 201596_x_at
#> result.15         1           1         0         1           1           0
#>                   0           1         0         0           1           1
#>                   0           1         0         0           1           1
#> result.3          0           1         0         0           1           1
#> result.9          0           0         0         0           1           0
#>                   1           1         0         1           1           0
#>                   0           1         0         1           1           1
#>                   1           1         1         1           1           0
#> result.13         1           1         1         1           1           1
#>                   0           1         0         0           1           1
#> result.14         1           1         0         0           1           1
#> result.9          0           0         0         1           0           0
#>                   1           1         0         1           1           0
#>                   1           1         0         1           1           0
#>                   1           0         1         1           1           1
#>           214612_x_at 210727_at 208829_at 212199_at 211513_s_at 207804_s_at
#> result.15           0         1         1         1           1           1
#>                     1         0         0         1           0           1
#>                     1         0         0         0           1           1
#> result.3            1         0         0         1           0           1
#> result.9            1         1         0         0           0           0
#>                     1         1         1         1           1           1
#>                     1         1         1         1           1           0
#>                     1         0         0         1           0           1
#> result.13           1         0         0         0           1           1
#>                     1         0         0         1           0           1
#> result.14           1         0         0         1           0           1
#> result.9            0         0         0         0           0           1
#>                     0         1         1         1           1           1
#>                     1         0         1         1           1           1
#>                     1         0         0         0           1           1
#>           216266_s_at 221768_at 203597_s_at 216830_at 208231_at 206201_s_at
#> result.15           1         1           1         0         0           1
#>                     0         0           1         0         1           1
#>                     0         0           1         0         1           1
#> result.3            0         0           1         0         1           1
#> result.9            0         0           1         1         1           0
#>                     1         1           1         0         0           1
#>                     0         0           1         0         1           1
#>                     0         1           1         1         1           1
#> result.13           0         0           1         0         1           1
#>                     0         0           1         0         1           1
#> result.14           0         0           1         0         1           1
#> result.9            0         0           0         1         0           0
#>                     1         1           1         0         0           1
#>                     1         1           1         0         0           1
#>                     0         0           1         0         1           1
#>           205317_s_at 218268_at 208994_s_at 204449_at 201712_s_at 221433_at
#> result.15           1         1           1         1           1         1
#>                     1         1           1         0           0         1
#>                     1         1           1         0           0         1
#> result.3            1         1           1         0           0         1
#> result.9            1         0           1         0           0         1
#>                     1         1           1         1           1         1
#>                     1         1           1         0           1         1
#>                     1         1           1         0           0         1
#> result.13           1         1           1         1           0         1
#>                     1         1           1         0           0         1
#> result.14           1         1           1         0           0         1
#> result.9            0         0           0         0           1         0
#>                     1         1           1         1           1         1
#>                     1         1           1         1           1         1
#>                     1         1           1         1           0         1
#>           218767_at 221789_x_at 204231_s_at 218834_s_at 204282_s_at 213638_at
#> result.15         1           0           0           1           0         1
#>                   0           0           1           0           1         1
#>                   1           0           1           1           1         1
#> result.3          0           1           1           0           1         1
#> result.9          1           0           1           1           1         0
#>                   1           0           0           0           1         1
#>                   1           0           1           0           1         1
#>                   0           1           1           1           1         1
#> result.13         1           1           1           0           1         1
#>                   0           1           1           0           1         1
#> result.14         0           1           1           0           1         1
#> result.9          0           0           0           0           0         0
#>                   1           0           0           1           0         1
#>                   1           1           0           1           0         1
#>                   1           1           1           1           1         1
#>           202707_at 201810_s_at 211095_at 203764_at 219582_at 214583_at
#> result.15         1           0         0         1         1         0
#>                   0           0         1         1         0         1
#>                   0           0         1         1         0         0
#> result.3          0           0         0         0         0         1
#> result.9          1           0         0         0         1         0
#>                   1           0         0         1         1         1
#>                   0           1         1         1         0         1
#>                   0           1         1         1         0         1
#> result.13         0           0         1         1         0         1
#>                   0           0         0         0         0         0
#> result.14         0           0         0         1         0         1
#> result.9          0           1         0         0         0         0
#>                   1           0         0         1         0         0
#>                   1           0         1         1         1         0
#>                   0           0         1         1         0         0
#>           220170_at 206058_at 207432_at 220490_at 202572_s_at 203548_s_at
#> result.15         1         0         1         1           1           1
#>                   1         1         0         1           1           0
#>                   1         1         0         1           1           0
#> result.3          1         1         0         1           1           0
#> result.9          1         0         1         0           0           1
#>                   1         1         1         1           1           0
#>                   1         1         0         1           1           0
#>                   1         1         1         1           1           0
#> result.13         1         1         0         1           1           0
#>                   1         1         0         1           1           0
#> result.14         1         1         0         1           1           0
#> result.9          0         1         0         0           0           0
#>                   1         0         1         1           1           1
#>                   1         1         1         1           1           1
#>                   1         1         0         1           1           0
#>           220496_at 212691_at 206111_at 206765_at 217469_at 202228_s_at
#> result.15         1         1         1         1         1           1
#>                   1         0         1         1         0           0
#>                   1         0         1         0         1           0
#> result.3          1         0         1         1         0           0
#> result.9          1         0         0         0         1           0
#>                   1         1         1         1         0           1
#>                   1         0         1         1         1           0
#>                   0         1         1         1         1           0
#> result.13         1         0         1         1         1           0
#>                   1         0         1         1         0           0
#> result.14         0         0         1         1         0           0
#> result.9          0         0         0         0         0           0
#>                   1         1         1         1         1           1
#>                   1         1         0         1         1           1
#>                   1         0         1         1         1           0
#>           210247_at 214943_s_at 218941_at 201373_at 202009_at 213335_s_at
#> result.15         1           1         0         1         1           1
#>                   0           0         1         1         0           1
#>                   0           1         1         1         0           1
#> result.3          0           0         0         1         0           0
#> result.9          0           0         0         1         1           1
#>                   1           1         0         1         1           1
#>                   0           1         1         1         1           1
#>                   1           0         1         1         1           1
#> result.13         1           1         1         1         0           1
#>                   0           0         0         1         0           0
#> result.14         1           0         1         1         0           0
#> result.9          0           0         0         0         0           0
#>                   1           1         0         1         1           1
#>                   1           1         0         1         1           1
#>                   1           1         1         1         0           1
#>           206921_at 206564_at 212059_s_at 216443_at 219662_at 1007_s_at
#> result.15         1         1           1         1         1         1
#>                   1         1           1         1         1         0
#>                   1         0           0         1         1         0
#> result.3          1         1           1         0         1         0
#> result.9          1         0           0         1         0         1
#>                   1         1           1         0         1         1
#>                   1         1           1         1         1         0
#>                   1         1           1         1         1         0
#> result.13         1         1           0         0         1         1
#>                   1         1           1         0         1         0
#> result.14         1         1           1         0         1         0
#> result.9          0         1           0         0         0         0
#>                   1         1           1         1         1         1
#>                   1         1           1         1         1         1
#>                   1         0           0         0         1         1
#>           213621_s_at 202481_at 221217_s_at 203260_at 215428_at 220867_s_at
#> result.15           1         1           1         1         1           1
#>                     0         1           0         1         0           1
#>                     0         1           0         1         0           1
#> result.3            0         1           0         1         0           0
#> result.9            1         0           0         1         0           0
#>                     1         1           1         1         1           1
#>                     0         1           0         1         0           1
#>                     1         1           0         0         1           1
#> result.13           0         1           0         1         0           1
#>                     0         1           0         1         0           0
#> result.14           0         1           0         1         0           0
#> result.9            0         1           0         1         0           0
#>                     1         1           1         1         1           1
#>                     1         1           1         1         1           0
#>                     0         1           0         1         0           1
#>           211426_x_at 221658_s_at AFFX-HUMRGE/M10098_3_at 212019_at 213500_at
#> result.15           1           0                       1         1         1
#>                     0           0                       1         1         0
#>                     0           1                       1         1         0
#> result.3            0           1                       1         1         0
#> result.9            1           1                       1         1         0
#>                     0           0                       1         1         0
#>                     0           1                       1         1         1
#>                     0           1                       1         1         1
#> result.13           0           1                       0         1         0
#>                     0           1                       1         1         0
#> result.14           0           1                       1         1         0
#> result.9            0           0                       0         0         0
#>                     1           0                       1         1         1
#>                     1           0                       1         1         1
#>                     0           1                       0         1         0
#>           221259_s_at 215012_at 207535_s_at 202118_s_at 204625_s_at 206540_at
#> result.15           1         1           1           1           0         1
#>                     1         1           1           1           0         0
#>                     1         1           1           1           0         1
#> result.3            0         1           1           1           0         0
#> result.9            1         1           0           0           1         0
#>                     0         1           1           1           0         0
#>                     1         1           1           1           0         1
#>                     1         1           1           1           1         0
#> result.13           1         1           1           1           0         1
#>                     1         1           1           1           0         0
#> result.14           0         0           1           1           0         0
#> result.9            0         0           0           0           0         0
#>                     1         1           1           1           0         1
#>                     1         1           1           1           0         1
#>                     1         1           1           1           0         1
#>           214608_s_at 203682_s_at 206203_at 200753_x_at 202518_at 207936_x_at
#> result.15           1           1         0           1         1           1
#>                     0           1         1           0         1           0
#>                     0           1         1           1         1           0
#> result.3            1           1         1           0         1           0
#> result.9            0           0         0           1         1           0
#>                     1           1         0           0         1           1
#>                     0           1         1           1         1           1
#>                     1           1         1           0         1           0
#> result.13           1           1         1           1         1           0
#>                     1           1         1           0         1           0
#> result.14           1           1         1           0         1           0
#> result.9            0           0         0           0         0           0
#>                     1           1         0           1         1           1
#>                     1           1         0           1         1           1
#>                     0           1         1           1         1           0
#>           215183_at 208069_x_at 204159_at 213965_s_at 221617_at 216958_s_at
#> result.15         1           1         1           0         1           1
#>                   1           0         0           0         1           1
#>                   1           1         0           0         1           1
#> result.3          1           0         0           0         1           1
#> result.9          1           1         1           0         1           0
#>                   1           1         0           0         1           1
#>                   1           1         0           0         1           0
#>                   0           0         0           1         1           1
#> result.13         1           1         0           0         1           1
#>                   1           0         0           0         1           1
#> result.14         1           0         0           0         1           1
#> result.9          0           1         0           0         0           1
#>                   1           1         1           0         1           1
#>                   1           1         1           0         1           1
#>                   1           1         0           0         1           1
#>           215762_at 214682_at 213605_s_at 212229_s_at 203492_x_at 218824_at
#> result.15         0         1           1           1           0         1
#>                   0         0           1           1           1         1
#>                   0         0           1           1           1         1
#> result.3          1         0           1           1           1         1
#> result.9          0         0           0           0           1         0
#>                   0         0           1           1           1         1
#>                   1         0           1           1           1         1
#>                   1         0           0           1           1         1
#> result.13         1         0           0           1           1         1
#>                   1         0           1           1           1         1
#> result.14         1         0           0           1           1         1
#> result.9          0         0           0           1           0         0
#>                   0         1           1           1           0         1
#>                   1         1           1           1           0         1
#>                   1         0           0           1           1         1
#>           220812_s_at 210872_x_at 213922_at 214159_at 209535_s_at  k
#> result.15           1           1         1         1           1  7
#>                     1           1         0         0           1  2
#>                     1           1         0         0           1  5
#> result.3            1           1         0         0           1  2
#> result.9            1           0         0         1           0 10
#>                     1           1         1         1           1  6
#>                     1           1         0         0           1  2
#>                     1           1         1         1           1  2
#> result.13           1           1         1         0           1  2
#>                     1           1         0         0           1  2
#> result.14           1           1         0         0           1  2
#> result.9            0           0         0         0           0  2
#>                     1           0         1         1           1  7
#>                     1           1         1         1           1  7
#>                     1           1         1         0           1  5
#>                                 rnkIndex     CrowD
#> result.15  0.02431213 262.79073        1       Inf
#>            0.09865128 230.56927        1       Inf
#>            0.03678756 252.92541        1 0.7816078
#> result.3   0.09704814  91.66167        2       Inf
#> result.9  -0.01429944 238.74222        2       Inf
#>            0.03088650 237.73015        2 0.7875784
#>            0.05287022 188.39536        2 0.5683238
#>            0.05556335 146.41825        2 0.4869773
#> result.13  0.08385136 132.88299        2 0.4031756
#>            0.09277432 127.17414        2 0.1793662
#> result.14  0.09667224 115.69393        2 0.1738992
#> result.9   0.09036216  58.64087        3       Inf
#>            0.01611480 215.25780        3       Inf
#>            0.02984614 213.94958        3 0.4159988
#>            0.03778117 156.74672        3 0.4073372
#> 
#> Slot "ParetoFront":
#> [[1]]
#>                   [,1]      [,2]
#> result.1   0.044079314  71.19605
#> result.2   0.026139667  17.30654
#> result.3   0.012209631 111.77052
#> result.4   0.021754619 118.08292
#> result.5   0.006081003  40.16175
#> result.6   0.036979925 148.63547
#> result.7   0.048008684  42.93976
#> result.8   0.033147918  75.26986
#> result.9   0.090362158  58.64087
#> result.10  0.021572026  84.70667
#> result.11 -0.014222997 112.22295
#> result.12  0.018121489   5.86589
#> result.13  0.015541257  94.55781
#> result.14  0.096672242 115.69393
#> 
#> [[2]]
#>                                  
#> result.3   0.0970481380  91.66167
#> result.15  0.0243121338 262.79073
#> result.6   0.0369799250 148.63547
#> result.13  0.0838513642 132.88299
#> result.14  0.0966722422 115.69393
#> result.9   0.0903621580  58.64087
#> result.9  -0.0142994429 238.74222
#> result.1   0.0691510166 129.91117
#> result.6   0.0134643264 209.73971
#> result.7   0.0480086837  42.93976
#> result.8   0.0001009192 149.45557
#> result.1   0.0440793139  71.19605
#> result.10  0.0086043263 130.42500
#> result.4   0.0217546192 118.08292
#> result.12  0.0345737519 102.28249
#> 
#> [[3]]
#>                                  
#> result.3   0.0970481380  91.66167
#> result.15  0.0243121338 262.79073
#> result.6   0.0369799250 148.63547
#> result.13  0.0838513642 132.88299
#>            0.0367875605 252.92541
#> result.14  0.0966722422 115.69393
#> result.9   0.0903621580  58.64087
#> result.9  -0.0142994429 238.74222
#> result.1   0.0691510166 129.91117
#>            0.0429419377 132.25239
#> result.6   0.0134643264 209.73971
#>            0.0223463738 156.12843
#>            0.0141982901 180.76370
#> result.8   0.0001009192 149.45557
#>            0.0585861250  60.70461
#> 
#> [[4]]
#>                                
#> result.3   0.09704814  91.66167
#> result.15  0.02431213 262.79073
#> result.6   0.03697993 148.63547
#>            0.03678756 252.92541
#> result.13  0.08385136 132.88299
#>            0.05556335 146.41825
#> result.14  0.09667224 115.69393
#> result.9   0.09036216  58.64087
#> result.9  -0.01429944 238.74222
#> result.1   0.06915102 129.91117
#> result.6   0.01346433 209.73971
#>            0.02234637 156.12843
#>            0.01419829 180.76370
#>            0.04052784 132.31614
#>            0.04294194 132.25239
#> 
#> [[5]]
#>                                
#> result.15  0.02431213 262.79073
#>            0.09865128 230.56927
#>            0.03678756 252.92541
#> result.3   0.09704814  91.66167
#> result.9  -0.01429944 238.74222
#>            0.03088650 237.73015
#>            0.05287022 188.39536
#>            0.05556335 146.41825
#> result.13  0.08385136 132.88299
#>            0.09277432 127.17414
#> result.14  0.09667224 115.69393
#> result.9   0.09036216  58.64087
#>            0.01611480 215.25780
#>            0.02984614 213.94958
#>            0.03778117 156.74672
#> 
#>